feat(v1alpha1): Introduce status path for initramfs
This commit introduces, similar to the kernel path, the status path for
the initram filesystem (CPIO archive).  This is used as status and as
part of the usual instantiation of a machine across existing VMM
solutions.  Simultaneously, the QEMU implementation of the machine
v1alpha1 interface is updataed to accomodate for this status path.
Unlike "rootfs", initramfs is a systematically defined path, whereas
rootfs is inteded to be a user-specified path which should ultimately be
resolved from different high-level locations.

Signed-off-by: Alexander Jung <alex@unikraft.io>