refactor(run): Add runner aliasing
This commit introduces aliasing to the `PackageManager` to enable users
to specifically set the PM of choice via the `--as` flag.  This keeps
the CLI consistent with other appearances of `--as`.

Simultaneously, the value of the `--as` flag is validated in `Pre`.  We
can unfortunately not use the general-purpose `EnumFlag`, since
injecting the flag in `New` will not resolve registered
`PackageManager`s.  As a result, we introduce a naive alternative.

Signed-off-by: Alexander Jung <alex@unikraft.io>