refactor(unikraft)!: Migrate default build directory into `.unikraft/`
To keep the project directory cleaner and unikraft-specific elements
within a contained space, this commit migrates the default build
directory from `build/` to `.unikraft/build/`.  Since the normal
`build/` directory is not hidden, it also risks a collision with a
user-provided directory of the same name whereas the hidden directory
`.unikraft` is less likely to have a collision.

Signed-off-by: Alexander Jung <alex@unikraft.io>