fix(run): Correctly handle cleaning up with `--rm`
This commit fixes an issue where the network was pre-emptively removed
before the invocation of the unikernel.  This is because the `defer`
clean up was migrated into the network parser.  Of course, at the end of
this function it is invoked.  Instead, it should be part of the main
method that handles the creation and start of the unikernel such that in
this lifecycle it is handled correctly.

Signed-off-by: Alexander Jung <alex@unikraft.io>