fix(exec): Remove stdin/stdout assumptions
This commit removes both the dependency to `iostreams` but also the
assumption that if stdin/stdout are not set that they should be set to
the configured stdin/stdout.  This can have undesired consequences when
spawning a process that has a stdin, rendering the implementing program
unusable since it is now tied to the exec'd program.  The same is true
for stdout, where the output can mangle the output of the implementing
program.  The use of `WithStdin`/`WithStdout`, are designed to be used
when redirection is desired and to provide the mechanism of how.

Signed-off-by: Alexander Jung <alex@unikraft.io>