fix: Better handling of directory sources
This commit fixes an issue where a directory can be specified as the
source.  This is useful in circumstances where the user is directly
manipulating the library's source code or it is not a public/upstream
microlibrary/component.  Setting the path will correctly resolve the
manifest's `DirectoryProvider`.

Additionally, skip "finding" and "pulling" components that have a path
that is the same as the source, which is the result of a
microlibrary/component located on disk.

Signed-off-by: Alexander Jung <alex@unikraft.io>