feat!: Adjust pulling for a particular platform
To accommodate for additional platforms, adjust both the `kraft run`
command and the underlying implementation for OCI package manager to
allow for specifying the platform during the `FetchImage` sequence
provided by the handler.

Additionally, the `WithPullArchitecture` and `WithPullPlatform`
`Package` options were adjusted from a slice of strings to simply a
string.  It does not make sense to pull more than one platform or
architecture as a general user of KraftKit.

Signed-off-by: Alexander Jung <alex@unikraft.io>