fix(run): Instantiate package manager is not kernel or project
This commit checks whether the runner-type has been pre-emptively set to
"project" or "kernel".  These runners do not require access to the
package manager which would otherwise be instantiated, slowing down the
instantiation process via `kraft run`.

In the future, it may make sense to simply embbed the user of the
package manager in the relevant runner.

Signed-off-by: Alexander Jung <alex@unikraft.io>