feat(platform): Introduce platform strategy iterator
This commit introduces a new method `NewMachineV1alpha1ServiceIterator`
which instantiates a `v1alpha1.MachineService`-compatible interface
which iterates over each representing method.  This is useful in
circumstances where a platform is not defined an each platform's method
can be called sequentially.  The first platform strategy to succeed is
returned.

Signed-off-by: Alexander Jung <alex@unikraft.io>