feat(oci): Add directory handler
This PR introduces a new OCI "directory" handler. This is used as a
fallback when no underlying client (containerd for now) is provided,
without any prerequisites. It stores all the images as raw layers and
manifests under the KRAFTKIT_RUNTIME_DIR. This handler enables users
to pull, run and save Unikraft OCI images out-of-the-box.

Signed-off-by: Luca Seritan <luca.seritan@gmail.com>
Signed-off-by: Alexander Jung <alex@unikraft.io>