fix: Correctly access the KConfigTree
This commit makes a notable change to the application's method
`KConfigTree` which a). correctly populates the values based on the
internal Unikraft make target `print-vars` and b). uses the Unikraft
core KConfigTree to access this tree since it derives the configuration.

An adjustment to the prototype of `component.Component` interface is
made to access context, this project-wide refactor is included.

Signed-off-by: Alexander Jung <alex@unikraft.io>