feat(initrd): Introduce `NewFromMapping`
NewFromMapping accepts a working directory base that represents
the relative path of the files to serialize which are provided by
a slice of mappings, each separated by the default delimeter
`:` which separates the path on the host and the path inside of
the initramfs.  A destination of the final archive is provided by
the positional argument output.

Signed-off-by: Alexander Jung <alex@unikraft.io>