feat(run): Dynamically serialize a provided path to a CPIO archive
This commit adjusts the functionality of the `-i|--initrd` flag such
that a path or file can be specified.  When a directory is provided to
the flag, it is dynamically serialized into a CPIO archive.  This eases
the use of `kraft run` without having to prepare an archive prior.

Additionally, the example of the command is updated, including overall
formatting, to include the usage of this flag adjustment.

Signed-off-by: Alexander Jung <alex@unikraft.io>