refactor(build)!: Deprecate `--fast` for `--no-fast`
This commit reverses the default functionality of `kraft build` such
that its invocation without any flags automatically uses the maximum
number of cores (aka "fast mode") when performing the build.  Not only
is now the build much faster by default, but the user must specifically
dictate that they do not want this functionality in a similar fashion
to other switches in `kraft build` via the new flag `--no-fast`.  The
use of maximum parallelization also respects the user's configuration.

Signed-off-by: Alexander Jung <alex@unikraft.io>