feat(build): Introduce `--all` flag and target selection menu
The previous default behavior of `kraft build` was to build all
platforms and architecture target combinations.  This leads to some
confusion in some cases.  Instead, provide a utility prompt menu (if
allowed by the user's configuration) to select the target to construct.
To resume the previous functionality, the user can use the newly
introduced `--all` flag.

Signed-off-by: Alexander Jung <alex@unikraft.io>