refactor!: Introduce `PlatformsByName` and `PlatformAliases`
This commit adjusts the `Platforms` method to return a slice and
introduces two helper methods `PlatformsByName` and `PlatformAliases`
which are used to handle the situation where a platform can have an
alias.  For example: "kvm" and "qemu" with the QEMU platform, and
"fc" and "firecracker" with Firecracker.

This API adjustment is propagated across the codebase in the same
commit.

Signed-off-by: Alexander Jung <alex@unikraft.io>