fix(plat): Handle platform name aliasing when parsing Kraftfiles
If a user has specified a known internal-to-Unikraft platform name
alias, handle the rewrite when the schema is transformed.  This
fixes an issue where a user may specify, for example, "kvm" or
"firecracker" under the `platform` attribute in a Kraftfile and the
kernel name would not be correctly set and therefore the build or run
would not perform correctly with `kraft`.

Signed-off-by: Alexander Jung <alex@unikraft.io>