refactor(build)!: Remove the preparation step
This commit totally removes the preparation step before the build (and
also removes the `--no-prepare` flag).  This is a redundant step as it
is technically accomplished through Unikraft's core build system and is
a spill-over from the earlier pykraft implementation which required a
separation between these steps.  It is no longer necessary.

Signed-off-by: Alexander Jung <alex@unikraft.io>