fix(menu): Fix hanging `kraft menu`
This commit fixes an issue where the Unikraft menuconfig system would
not open and the command `kraft menu` would hang.  This issue was
introduced in ba4153ba when the default functionality of the `exec`
package removed IOStreams as the command's std{in,out,err}.  This change
is still valid, the implementing method should decide what these values
are and in this commit, we do this with the invocation of the
menuconfig.

Signed-off-by: Alexander Jung <alex@unikraft.io>