refactor(manifest): Improve sequential `Catalog`-call caching
This commit adjusts the internal `ManifestManager` to better handle
temporary in-memory caching to speed up sequential calls to the
`Catalog` whilst still respecting the use of the `QueryOption` for using
cache.  When `Update` is called, the in-memory cache will be updated.
Otherwise, as soon as at least 1 update is made, we can use this
in-memory representation of the `ManifestIndex` to search for packages.
This reduces the number of calls to the remote manifest index and
reduces the execution time.

Signed-off-by: Alexander Jung <alex@unikraft.io>