feat(run): Introduce volume flag and attach to machine
This commit introduces a new flag slice `-v|--volume` which allows the
user to attach a specified source to a given destination within the
instantiated unikernel machine instance.  To make this change, the
driver is automatically detected, looked up, and the machine
specification is adjusted appropriately based on compatibility and
capability.

Signed-off-by: Alexander Jung <alex@unikraft.io>