feat(oci): Implement directory handler pushing
There are currently 2 OCI specifications used. The `containerregistry`
implementation and the `opencontainers` implementation.
`containerregistry` was used to interface with registries whilst the
`opencontainers` specification was used to save data.

As part of pulling, the translation from `containerregistry` to
`opencontainers` was already implemented. The reverse is needed for
pushing.

This commit implements two interfaces for interacting with
`containerregistry` functions related to pushing.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>