feat(tools): Introduce github-action tool
This tool is designed to be used in a CI/CD context on GitHub in an
actions workflow and aims to simplify the entire process of using
KraftKit by offering several high-level abstractions that are useful
in this context.

The ultimate use case is for this new GitHub Action is when a user
has a Kraftfile in their repository and to get started building,
packaging and publish the image by simply introducing a small, well-
defined uses step. The following is an example `uses` step which
this tool ultimately aims to offer:

```yaml
- uses: unikraft/kraftkit@stable
  with:
    workdir: path/to/project
    kraftfile: path/to/Kraftfile
    arch: x86_64
    plat: qemu
    push: false
    output: oci://ghcr.io/username/repo/unikernel:latest
    kconfig: true
```

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Co-authored-by: Alexander Jung <alex@unikraft.io>