feat: Upgrade QEMU from v7.1.0 to v7.2.4
This is the latest version of the v7 series.  At the same time, set the
QMP version such that it is reflected via the import.  We are setting
`v7alpha2` here instead of `v7` because the QMP implementation is
incomplete.  If it is fully populated we can rename to `v7`,
representing a complete API that can then be referenced.

Signed-off-by: Alexander Jung <alex@unikraft.io>