fix(github-action): Re-use project instantiation and target selection
Parts of this code was copied verbatim instead of adapted, resulting in
duplicate code within the action.  This commit simply removes elements
(instantiation of a project and selection of a target) such that this
values are computed only once.  Additionally, in some cases, the value
were miscalculated (such as when a custom Kraftfile is specified as a
string).

Signed-off-by: Alexander Jung <alex@unikraft.io>