fix(store): Attempt once before starting retry loop
Sometimes accessing the files in the runtime directory results in a
block as the program assumes that the directory is locked and retries
in a loop to access it for 5 seconds.

Most of the time, this happens because of no permissions.

This tries to run once and then check the error message for permission
problems.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>