feat: Add `kraftld`
This implements a `kraftld` linker shim acting as a gcc-style linker.

It accepts the same arguments and forwards them to `UK_LDFLAGS`.
It removes flags that are known to be incompatible.
Furthermore, it sanitizes relative paths for the Unikraft build system.

`kraftld` also extracts any file dependencies for `UK_LDEPS`.
Lastly, it parses `-o <OUTPUT>` and copies the KraftKit output image.
Currently, it assumes the `default_qemu-x86_64`.
Future versions of KraftKit will support an output option.
`kraftld` will then adapt that option instead.

`kraftld` is currently targeting the Rust compiler.
Other compilers may or may not work as well.

Signed-off-by: Martin Kröning <martin.kroening@eonerc.rwth-aachen.de>
GitHub-Closes: #612