feat(dockerfile-llb-frontend): Add the initial LLB frontend code.
This code results in a working LLB plugin that can then be consumed with
docker to build a unikraft app.

Includes unit tests for the core build part, end-to-end tests suites for
BuildKit and Docker, and a comprehensive readme.

The end to end tests build the app-helloworld.

The BuildKit test suite is skipped by default as the docker one
does the same test of correctness. Its usefulness is mostly for local
debugging.

Add a GitHub Action to build and push the kraftkit.sh/llb image.

GitHub-Fixes: #707
GitHub-Fixes: #708

Signed-off-by: Jakub Ciolek <jakub@ciolek.dev>