feat(app): Allow instantiating kraftfiles from byte slices.
When working with BuildKit, the LLB plugin doesn't have
a direct access to the client-local filesystem.

BuildKit provides files on a request-response basis and
the interface is a slice of bytes.

Allow instantiating kraftfiles from raw byte slices
to decouple the process from trying to directly
interface with the filesystem.

Signed-off-by: Jakub Ciolek <jakub@ciolek.dev>