feat(run): Enable running LLB-built docker images
WIP

This introduces a `dockerImage` package to run docker images
containing a unikernel built with the LLB plugin.

The hello world target runs successfully.

A new package manager, "docker", has been added.
If the target OCI image is missing, this package manager
searches for it.

There's an ongoing discussion about how to handle image references,
especially when an image could be in both local/remote OCI storage
and local docker storage.

Currently, the process first checks the OCI storage and then Docker.

In case of conflicts, the user might need to choose.

To-Do:

- Fix targets that don't build due to missing dependencies or artifact
  naming issues.
- Provide architecture/platform info via image, as Docker doesn't
  support non-standard OS like qemu.
- Consider support for remote docker registries.
- Refactor some interfaces; we could make some of them smaller and avoid
  the "not implemented" errors.
- Explore testing approaches for this functionality.

Signed-off-by: Jakub Ciolek <jakub@ciolek.dev>