feat(run): Detail 'memory' argument
The 'memory' argument can be specified in multiple formats.
The units of measurement can be: K/M/G/T which are multiples of 1000.
Or they can be: Ki/Mi/Gi/Ti which are multiples of 1024.
Because this is Kubernetes syntax, it is not really known to other
people except when documented.

Also switch the default value from M to Mi for better performance.

GitHub-Fixes: #773

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>