fix: Restore output writer after calling processtree
This is a temporary workaround for incorporating multiple processes in
a command. After calling processtree the original output writer is lost
so writing will no longer work. To fix this we temporarily save it
beforehand.

A proper fix would ensure in the tui package that this writer is
preserved. Thankfully, this is the only place where it manifests right
now.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>