feat(runu): handle CRI sandbox for Kubernetes
Runs the process defined in the container spec instead of QEMU if the
container is identified as a CRI sandbox container ("pause" container).

Annotates the container config with "org.unikraft.kernel" if the
container is an actual unikernel, to allow libmocktainer to tell
unikernel containers apart from sandbox containers.

Adds a "--all" flag to the "kill" subcommand, for backwards compatibility
with (cri-)containerd.

Adds a "ps" subcommand. Although this command isn't part of the OCI
runtime specification, it is invoked through CRI.

Signed-off-by: Antoine Cotten <antoine@unikraft.io>