refactor!: Remove unused target attribute
This commit removes the field method "format" from the target
specification.  The usage was intended to be part of the Kraftfile
specification, wherein a user would specify the format of the target
via:

```yaml
targets:
- name: example
  architecture: x86_64
  platform: qemu
  format: oci
```

However, this was never realised and introduces some complications for
now that are not resolved, specifically to do with the name of the
target which conflicts with internal usage (by both the `unikraft/`
package and the relevant package manager).  In this commit, refactor
the `kraft pkg` subcommand to ignore this attribute

Signed-off-by: Alexander Jung <alex@unikraft.io>