refactor: Use dependency injection for configuration passing
Switch from using the context based config accessor
towards dependency injection.

Changes the architecture a bit so that config data flow is more
explicit.

Left a few todo's, but overall this simplifies reasoning about config
handling. Should make unit/integration testing easier.
Improves clarity when debugging.

The generic stuff is still around, but we can think whether the added
abstractions are worth it and whether using a simpler way would be
better.

Another architecture decision point is on what type of the param shall
we inject. Perhaps an interface would be better.

The test suite results are equivalent to 2828af22. The app-helloworld
target builds and runs.

Signed-off-by: Jakub Ciolek <jakub@ciolek.dev>