fix: Use implicit knowledge about the default manifest to speed up init
In most cases, the user will have the default manifest location.  It is
not necessary to check if this is compatible since it is inherently
compatible.  Instead of performing this check, which ultimately makes a
remote connection call, add the manifest and do not run `IsCompatible`.
To make this possible, export the constant representing default
location.

Signed-off-by: Alexander Jung <alex@unikraft.io>