refactor(oci)!: Remove autosave
This commit removes the "auto-save" ability from the OCI image
abstraction.  This is done for two reasons: a). it is not used anywhere
in the KraftKit codebase and b). it has an affect on containerd's
garbage collector.  Image components must be saved in a particular order
in order to ensure they are not removed by the garbage collector.  To
prevent this, removing autosave reduces complexity and ensures that that
during the image's `Save` method, items are performed in the correct
order.

Signed-off-by: Alexander Jung <alex@unikraft.io>