refactor(oci): Implement `ResolveManifest` and `DeleteManifest` handlers
This commit reworks the OCI package's `Delete` method.  In its previous
form, it assumed the directory implementation's handler which is
incorrect since the OCI package manager relies on an underlying
implementation for storing the artifacts of the OCI image spec.  To
accomplish this, containerd and the directory handler were updated with
relevant methods `ResolveManifest` and `DeleteManifest` interface
methods.

Signed-off-by: Alexander Jung <alex@unikraft.io>