fix(manifest): Improve initialization speed
This commit uses inherent knowledge about KraftKit to speed up the
initialization of the manifest package manager by preventing a remote
checks via an update.  If sources have been added to the manifest, they
are likely valid.  Instead, run the update only if a request to update
is made.  When no sources in the internal manifest cache, then we
iterate over each manifest in the config and save appropriately.

Signed-off-by: Alexander Jung <alex@unikraft.io>