fix(oci): Correctly handle saving the manifest
This commit fixes an issue where the composition of a manifest is not
saved correctly by containerd.  Whilst the change occurs at a level
higher than the containerd handler, the order here is important since
this is recognised by containerd and its garbage collector.  Ensure that
the manifest is first saved and then its entities such that the
previously unreferrenced entities are known before they are saved.

Signed-off-by: Alexander Jung <alex@unikraft.io>