feat(handler): Introduce `SaveDescriptor` method
`SaveDescriptor` accepts an optional name reference which represents
descriptor (but this is not always necessary and can be left blank if
the descriptor is unnamed, e.g. an untagged config, a layer, etc) as
well as an io.Reader which is prepared to pass in the byte slice of
the descriptor.  An optional progress method callback can be provided
which is used to deliver the progress of writing the descriptor by
the implementing method.

Signed-off-by: Alexander Jung <alex@unikraft.io>