refactor(oci):! Implement and use OCI Image Indexes for mutli-images
This large refactor reworks the OCI package manager to assume the use of
the OCI image specification's index such that we can represent multiple
manifests within a single canonical referencable name.  This means that
we can build multiple unikernels with the same name.  This makes sense
for multi-target projects which build the same application but target
different hardware/platform vendors.

The structure of the index can be found at [0] and in this
implementation we itemized each manifest in the index in the same way.
When we use the `Catalog` method now, we are able to query a remote
registry based on the selected OS (platform) and architecture as well as
any embedded KConfig options (if they are embedded).

Use cases include selecting appropriately based on the host's
architecture as well as feature selection via embedded KConfig options.

Notable additional changes include:

- Removal of `FetchImage` which was replaced by `PullDigest`;
- Removal of `PushImage` which was replaced by `PushDescriptor`;
- Removal of `ResolveImage` which was replaced by `DigestExists`;
- Reworking the `Catalog` structure to parse indexes containing
  manifests; and,
- Introduction of `NewIndex`, `NewIndexFromSpec` and `NewIndexFromRef`
  which are used to populate an internal index representation from
  different sources.

[0]: https://github.com/opencontainers/image-spec/blob/main/image-index.md

Signed-off-by: Alexander Jung <alex@unikraft.io>