feat(internal): Import file locking packages verbatim
This commit copies in file locking packages from Go's internal
source[0] code verbatim.  The only adjustment made to the packages
are the removal of test cases which relied on additional imports
as well as changing the import path.

These packages are to be used when reading and writing files that
are performed asynchronously.

Finally, these files were passed through `gofumpt` to make their
syntax compliant with KraftKit's style guide.

[0]: https://github.com/golang/go/tree/go1.21.3/src/cmd/go/internal/lockedfile

Signed-off-by: Alexander Jung <alex@unikraft.io>