feat: Introduce ability to overwrite or merge during packaging
This commit introduces a "merge strategy" system and allows the user to
decide during the invocation of `kraft pkg` either by CLI or via TUI
prompt whether to overwrite or merge into an existing package.  For now
the OCI package manager package implements this and enables users to
construct unikernels sequentially whilst still retaining the same image
name.  Further, the user can pull an existing image and add new
kernels.

To make this possible, the packmanager package was updated to introduce
a native option `PackMergeStrategy` along with 3 merge strategies:

- StrategyExit: Exit on existing/conflict;
- StrategyMerge: Merge new targets with an existing package;
- StrategyOverwrite: Totally remove the existing package and set the
  contents to the new targets.

Additionally, the OCI package was updated to accommodate for this
functionality and iterates across each existing manifest and checks the
the metadata (platform, architecture and set KConfig options) to
determine whether to overwrite.

Signed-off-by: Alexander Jung <alex@unikraft.io>