fix(oci): Speed up package manager initialization
Use internal KraftKit knowledge of the fact that the config often lists
the well-known path of the Manifest package manager's remote index.
This is obviously not an OCI image registry so we can safely skip it.
Doing this speeds up the kraft CLI and the instantiation of the OCI
Package Manager in general by a noticable amount, especially with
limited internet connectivity (as none is subsequently required).

Signed-off-by: Alexander Jung <alex@unikraft.io>