fix(pkg): Add `--force` flag to sourcing and prevent check on unsourcing
This commit fixes an issue whereby the user will have the compatibility
of a source checked before the source is saved to config.  To prevent
this check, a new `kraft pkg source --force` flag is added which
prevents this check.

Inversely, when a source is being removed from `kraft pkg unsource`, its
compatibility does not need to be checked because it is being removed,
not only is this less computation but if the source was manually added
by hand and it was invalid, it would be impossible to remove.

This commit also updates relevant tests which are impacted by this
change.  Specifically, tests that use a obviously fake source for
testing purposes, the `--force` command is now utilized to ignore what
would otherwise error.

Signed-off-by: Alexander Jung <alex@unikraft.io>