fix(workflows): Do not match citations on commit messages
This commit adds a fix to the regex pattern to also accept citations
that are longer than 74 characters.

There is still one outstanding issue with this. If a commit contains
lines over 74 characters and also citations longer than 74 characters it
will still match. This case is rare, and probably that commit needs
special attention anyways.

Signed-off-by: Cezar Craciunoiu <cezar@unikraft.io>