fix(pull): Use host plat/arch and rename flag `--all`
This commit first names the flag `--all-versions` to `--all`.  This
doesn't affect existing functionality since this flag was unused.
Instead, this flag is now used to determine whether to pull "everything"
from the selected catalog based on the provided name.  When unset, and
additionally when `--arch` and `--plat` are unset, the host is queried
for its architecture and platform, as a usability improvement.

Signed-off-by: Alexander Jung <alex@unikraft.io>