fix(oci): Handle case where remote and local digests exist
In the scenario where a user has already pushed the same manifest to a
remote registry and still has the manifest locally, the `Catalog` method
would return both entries.  Instead, use the platform checksum as a
unique index in a map to prevent duplicates.  The local will be
preferred since this is queried last.

Signed-off-by: Alexander Jung <alex@unikraft.io>