refactor(pkg): Rework packaging into characterizable packagers
This commit reworks the `kraft pkg` subcommand to accept different
packaging implementations.  By supplying an interface with the
pre-determination method `Packagable`, a candidate packaging mode can be
selected and run.  This paves the way for packaging inputs based on
different contexts.

For now identical functionality is of the `kraft pkg` remains.

Signed-off-by: Alexander Jung <alex@unikraft.io>