refactor(handler)!: Use a single digests directory and symlinks as tags
This internal refactor re-works how digests are saved by the
`DirectoryHandler` as well as how indexes are represented.  Since
manifests, configs, layers and in fact image indexes are represented by
their cheksum (digest), these are migratead to a common `digests` folder.

To accommodate for named aliases (tags), these are represented by the
`indexes` directory and symbolic links are used to point to the relevant
digest.

```
/
|-- digests/
|   `-- sha256/
|       |-- 119a272c8275e03443e3953fda777ed12a6614a27c...
|       |-- 3b5c8ab1d8706a0e3578e1e45feecdbd9bfe27da73...
|       |-- 7b2f9aeddaf22f836b7a5f0f2b052f66f348e13040...
|       `-- a337184ac4dd01ae24435eb81e206c190add120931...
`-- indexes/
    `-- unikraft.org/
        `-- helloworld/
            `-- latest -> /digests/sha256/119a272c8275...
```

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Co-authored-by: Alexander Jung <alex@unikraft.io>