fix: Do not `load` github-actions image, do not pre-pull
This commit fixes the GitHub Actions workflow for the github-actions
program.  Previously, the integration would not be properly run, in fact
it was running what was already in the registry.  Solve this by not
"load"ing the image (this unintuitive name does the opposite of what we
want) so that we can use the image locally.

Additionally, do not run the pull step in the action; this will be done
if the image is not present by the run step.

Signed-off-by: Alexander Jung <alex@unikraft.io>