refactor!: Update `cmdfactory.Run` interface prototype
This commit changes `cmdfactory`'s `Run` interface method to exclude the
`*cobra.Command` structure in favour of `context.Context`.  In all cases
for `kraft`, `cmd.Context()` was the only information used from the
previous structure.  In all cases for `runu`, light rework was made to
migrate access to flags to be performed within `Pre`.

This refactor allows the `Run` method to be re-used outside of the
context of the `cobra.Command`/`cmdfactory` lifecycle.

Signed-off-by: Alexander Jung <alex@unikraft.io>