fix(run): Better handling of retrieving and detecting runtime package
This commit reworks the lookup of the specified package for the
"package" runner.  First, the lookup via the `Catalog` method is wrapped
with the processtree to enable a loading dial whilst the package is
looked up.  Additionally, if multiple packages are detected and when the
`--arch` and `--plat` flags are unset, use the host to determine its
value.  If a package with those values are found, use this package.

Signed-off-by: Alexander Jung <alex@unikraft.io>