feat(list): Reduce output and provide `--all` flag to extend it
This commit reworks the query to the `Catalog` method such that the
host's platform and architecture are incorporated by default.  This
reduces the output, especially with the OCI package manager's catalog
which has an increased list of applications but these vary by
architecture and platform.

To show everything, however, introduce an `--all` flag.  This means that
if you want to see OCI applications in the remote registry which are
applicable to your host, you simply invoke:

```console
kraft pkg ls --apps --update
```

To view all variants of the same application, now include the `--all`
flag:

```console
kraft pkg ls --apps --update --all
```

Signed-off-by: Alexander Jung <alex@unikraft.io>