refactor(pkg): Migrate the procedures in `Run` to `Pkg`
Functionally, this does not change `kraft pkg` and how it works.
However, it allows the exposed `Pkg` method (which can be called by
other CLI programas or internally within KraftKit) to receive the list
of packages on successful result.  This is made possible due to the
modification of the `Pkg` prototype.

Signed-off-by: Alexander Jung <alex@unikraft.io>