feat(utils): Introduce common kraft CLI util and `BuildRootfs`
This package is intended to store common utility methods which can
be used across different kraft subcommands.  For now, simply
introduce the package with a common method, `BuildRootfs`, which
will be later used across relevant subcommands.

The `BuildRootfs` command is also updated to utilze the newly exposed
`WithArchitecture` method which is used to build an appropriate file-
system depending on the provided target(s).

Signed-off-by: Alexander Jung <alex@unikraft.io>